+++ SubEthaEdit-Mac/DNDArrayController.m
@@ -50,7 +50,7 @@ NSString *MovedRowsType = @"MOVED_ROWS_TYPE";
 
 - (NSDragOperation)tableView:(NSTableView*)tv
 				validateDrop:(id <NSDraggingInfo>)info
-				 proposedRow:(int)row
+				 proposedRow:(NSInteger)row
 	   proposedDropOperation:(NSTableViewDropOperation)op
 {
     
@@ -71,7 +71,7 @@ NSString *MovedRowsType = @"MOVED_ROWS_TYPE";
 
 - (BOOL)tableView:(NSTableView*)tv
 	   acceptDrop:(id <NSDraggingInfo>)info
-			  row:(int)row
+			  row:(NSInteger)row
 	dropOperation:(NSTableViewDropOperation)op
 {
     if (row < 0)
@@ -107,15 +107,15 @@ NSString *MovedRowsType = @"MOVED_ROWS_TYPE";
 
 
 -(void) moveObjectsInArrangedObjectsFromIndexes:(NSIndexSet*)indexSet
-										toIndex:(unsigned int)insertIndex
+										toIndex:(NSUInteger)insertIndex
 {
 	
     NSArray		*objects = [self arrangedObjects];
-	int			index = [indexSet lastIndex];
+	NSInteger			index = [indexSet lastIndex];
 	
-    int			aboveInsertIndexCount = 0;
+    NSInteger			aboveInsertIndexCount = 0;
     id			object;
-    int			removeIndex;
+    NSInteger			removeIndex;
 	
     while (NSNotFound != index)
 	{
